dadas

+++ labs/RayTracer/src/Figure.cpp
@@ -7,9 +7,9 @@ Figure::Figure()
 
 void Figure::read(tinyxml2::XMLElement* element) {
 
-    reflexion = atoi(element->Attribute("hasReflexion"));
-    kspec = atoi(element->Attribute("kspec"));
-    refraction = atoi(element->Attribute("hasRefraction"));
+    reflexion = atof(element->Attribute("hasReflexion"));
+    kspec = atof(element->Attribute("kspec"));
+    refraction = atof(element->Attribute("hasRefraction"));
     color = Vector(atof(element->Attribute("r")), atof(element->Attribute("g")), atof(element->Attribute("b")));
 }
