+++ labs/RayTracer/src/Vector.cpp
@@ -125,7 +125,7 @@ Vector Vector::vectorProduct(Vector v)
         return Vector((y * v.z) - (z * v.y),(z * v.x) - (x * v.z),(x * v.y) - (y * v.x));
 }
 
-Vector Vector::AddScalar(int scalar, Vector v){
+Vector Vector::AddScalar(float scalar, Vector v){
     Vector v2 = Vector(scalar*x + v.x, scalar*y + v.y, scalar*z + v.z);
     return Vector(scalar*x + v.x, scalar*y + v.y, scalar*z + v.z);
 }
